added server deletion if you are creator of a server and fixed channel messages which weren't deleted after its deletion

+++ models/channel.js
@@ -9,14 +9,15 @@ const channelSchema = new Schema({
 	messages: [{ type: Schema.Types.ObjectId, ref: "Message" }],
 });
 
-const Channel = mongoose.model("Channel", channelSchema);
 channelSchema.post("findOneAndDelete", async (channel) => {
-	if (channel.messages && channel.messages.length) {
+	if (channel && channel.messages && channel.messages.length) {
 		await Message.deleteMany({
 			_id: {
-				$in: Channel.messages,
+				$in: channel.messages,
 			},
 		});
 	}
 });
+const Channel = mongoose.model("Channel", channelSchema);
+
 module.exports = Channel;
